Refactor: Change project build system to poetry
