+++ gear_guardian/gear_guardian_gui.py
@@ -4,37 +4,49 @@ import os
 import argparse
 import json
 import cv2
-import tensorflow as tf
-from utils.utils import get_yolo_boxes, makedirs
-from utils.bbox import draw_boxes
+import tensorflow.compat.v1 as tf
+from tensorflow.compat.v1 import ConfigProto
+from tensorflow.compat.v1 import InteractiveSession
 from tensorflow.keras.models import load_model
 from tqdm import tqdm
 import numpy as np
-from object_tracking.application_util import preprocessing
-from object_tracking.deep_sort import nn_matching
-from object_tracking.deep_sort.detection import Detection
-from object_tracking.deep_sort.tracker import Tracker
-from object_tracking.application_util import generate_detections as gdet
-from utils.bbox import draw_box_with_id
+
+from .utils.utils import get_yolo_boxes, makedirs
+from .utils.bbox import draw_boxes
+from .object_tracking.application_util import preprocessing
+from .object_tracking.deep_sort import nn_matching
+from .object_tracking.deep_sort.detection import Detection
+from .object_tracking.deep_sort.tracker import Tracker
+from .object_tracking.application_util import generate_detections as gdet
+from .utils.bbox import draw_box_with_id
+from .app import App
 
 import warnings
 warnings.filterwarnings("ignore")
 
-from tensorflow.compat.v1 import ConfigProto
-from tensorflow.compat.v1 import InteractiveSession
+
+from multiprocessing import Queue
 
 config = tf.ConfigProto()
 config.gpu_options.allow_growth = True
 tf.keras.backend.set_session(tf.Session(config=config))
 
-def _main_(args):
-    config_path = args.conf
-    num_cam = int(args.count)
+def main():
+    # argparser.add_argument('-c', '--conf', help='path to configuration file')
+    # argparser.add_argument('-n', '--count', help='number of cameras')
+
+    app = App()
+    # config_path = args.conf
+    # num_cam = int(args.count)
+    config_path = "/home/pablo/WattHack-CopelHackaton/watthack-gear-guardian/config.json"
+    num_cam = 1
 
     with open(config_path) as config_buffer:
         config = json.load(config_buffer)
 
-    net_h, net_w = 416, 416
+    # makedirs(output_path)
+
+    net_h, net_w = 416, 416  # a multiple of 32, the smaller the faster
     obj_thresh, nms_thresh = 0.5, 0.45
 
     os.environ['CUDA_VISIBLE_DEVICES'] = config['train']['gpus']
@@ -61,6 +73,8 @@ def _main_(args):
 
     batch_size = num_cam
     images = []
+    values = []
+    messages = []
     while True:
         for i in range(num_cam):
             ret_val, image = video_readers[i].read()
@@ -72,14 +86,17 @@ def _main_(args):
                                          nms_thresh)
 
             for i in range(len(images)):
-                boxs = [[box1.xmin,box1.ymin,box1.xmax-box1.xmin, box1.ymax-box1.ymin] for box1 in batch_boxes[i]]
+                boxs = [[box1.xmin, box1.ymin, box1.xmax - box1.xmin, box1.ymax - box1.ymin] for box1 in batch_boxes[i]]
                 features = encoder(images[i], boxs)
-
+                message = ""
+                # print(features)
+                # score to 1.0 here).
                 detections = []
                 for j in range(len(boxs)):
                     label = batch_boxes[i][j].label
-                    detections.append(Detection(boxs[j], batch_boxes[i][j].c, features[j],label))
+                    detections.append(Detection(boxs[j], batch_boxes[i][j].c, features[j], label))
 
+                # Call the tracker
                 trackers[i].predict()
                 trackers[i].update(detections)
 
@@ -87,6 +104,8 @@ def _main_(args):
                 n_with_helmet = 0
 
                 for track in trackers[i].tracks:
+                    message+=track.message
+                    track.message=""
                     if not track.is_confirmed() or track.time_since_update > 1:
                         continue
                     if track.label == 2:
@@ -94,27 +113,22 @@ def _main_(args):
                     if track.label == 1:
                         n_with_helmet += 1
                     bbox = track.to_tlbr()
-                    # print(track.track_id,"+",track.label)
                     draw_box_with_id(images[i], bbox, track.track_id, track.label, config['model']['labels'])
 
-                # for det in detections:
-                #     print(det.label)
-                #     bbox = det.to_tlbr()
-                #     cv2.rectangle(images[i], (int(bbox[0]), int(bbox[1])), (int(bbox[2]), int(bbox[3])), (255, 0, 0), 2)
-                
-                print("CAM "+str(i))
-                print("Pessoas sem capacete  = " + str(n_without_helmet))
-                print("Pessoas com capacete  = " + str(n_with_helmet))
-                cv2.imshow('Cam'+str(i), images[i])
+                messages.append(message)
+                values.append(n_with_helmet)
+                values.append(n_without_helmet)
+            app.update(images, values,messages)
+            values = []
             images = []
+            messages = []
         if cv2.waitKey(1) == 27:
-            break  # esc to quit
+            break
     cv2.destroyAllWindows()
+
 
 if __name__ == '__main__':
     argparser = argparse.ArgumentParser(description='Predict with a trained yolo model')
-    argparser.add_argument('-c', '--conf', help='path to configuration file')
-    argparser.add_argument('-n', '--count', help='number of cameras')
 
     args = argparser.parse_args()
     _main_(args)
