+++ gear_guardian/object_tracking/application_util/freeze_model.py
@@ -1,6 +1,7 @@
 # vim: expandtab:ts=4:sw=4
 import argparse
-import tensorflow as tf
+
+import tensorflow.compat.v1 as tf
 import tensorflow.contrib.slim as slim
