+++ gear_guardian/object_tracking/application_util/generate_detections.py
@@ -4,7 +4,7 @@ import errno
 import argparse
 import numpy as np
 import cv2
-import tensorflow as tf
+import tensorflow.compat.v1 as tf
 
 
 def _run_in_batches(f, data_dict, out, batch_size):
