+++ gear_guardian/object_tracking/application_util/image_viewer.py
