+++ gear_guardian/object_tracking/application_util/visualization.py
