+++ gear_guardian/object_tracking/deep_sort/detection.py
