+++ gear_guardian/utils/bbox.py
