+++ gear_guardian/utils/colors.py
