+++ gear_guardian/utils/image.py
