+++ pyproject.toml
@@ -7,12 +7,20 @@ license = "MIT"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.8,<3.11"
+tensorflow-io-gcs-filesystem = { version = "0.27.0"}
+tensorflow = "2.12.0"
+opencv-python = "^4.8.1.78"
+tqdm = "^4.66.1"
+sklearn = "^0.0.post9"
+scikit-learn = "^1.3.1"
+pillow = "^10.0.1"
+
 
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-gear_guardian_cli = 'gear_guardian_cli:main'
-gear_guardian = 'gear_guardian_gui:main'
+gear_guardian_cli = 'gear_guardian.gear_guardian_cli:main'
+gear_guardian = 'gear_guardian.gear_guardian_gui:main'
